Deleted login and signup to implement upsertuser
