+++ src/App.js
@@ -1,7 +1,6 @@
 import { GlobalStyle } from "../src/styles/globalStyles";
 import { BrowserRouter, Switch, Route } from "react-router-dom";
-import Login from "./routes/Login";
-import SignUp from "./routes/SignUp";
+import UpsertUser from "./routes/UpsertUser";
 import Balances from "./routes/Balances";
 import AddEntry from "./routes/AddEntry";
 
@@ -10,10 +9,14 @@ export default function App() {
     <BrowserRouter>
       <GlobalStyle />
       <Switch>
-        <Route path="/" component={Login} exact />
-        <Route path="/signup" component={SignUp} exact />
-        <Route path="/balances" component={Balances} exact />
-        <Route path="/balances/:entryType" component={AddEntry} exact />
+        <Route path="/" component={UpsertUser} exact />
+        <Route path="/:userStatus" component={UpsertUser} exact />
+        <Route path="/:userStatus/:user" component={Balances} exact />
+        <Route
+          path="/:userStatus/:user/:entryType"
+          component={AddEntry}
+          exact
+        />
       </Switch>
     </BrowserRouter>
   );
