+++ src/routes/AddEntry.js
@@ -12,7 +12,7 @@ export default function AddEntry() {
   const [description, setDescription] = useState("");
   const history = useHistory();
   let entry;
-  const { entryType } = useParams();
+  const { entryType, user, userStatus } = useParams();
 
   if (entryType === "addcredit") {
     entry = {
@@ -62,7 +62,7 @@ export default function AddEntry() {
     <Body>
       <Header>
         {entry.name}
-        <Link to="/balances">
+        <Link to={`/balances/${user}`}>
           <RiArrowGoBackFill style={{ color: "white" }} />
         </Link>
       </Header>
@@ -72,18 +72,21 @@ export default function AddEntry() {
           placeholder="Valor"
           value={value}
           onChange={(e) => setValue(e.target.value)}
+          show={userStatus === "signup" ? false : true}
         />
         <Input
           type="text"
           placeholder="Descrição"
           value={description}
           onChange={(e) => setDescription(e.target.value)}
+          show={userStatus === "signup" ? false : true}
         />
         <Input
           type="submit"
           value="Salvar entrada"
           onClick={(e) => postNewEntry(e)}
           button={true}
+          show={userStatus === "signup" ? false : true}
         />
       </form>
     </Body>
